Crate hydro_lang

Hydro is a high-level distributed programming framework for Rust. Hydro can help you quickly write scalable distributed services that are correct by construction. Much like Rust helps with memory safety, Hydro helps with distributed safety.

The core Hydro API involves live collections, which represent asynchronously updated sources of data such as incoming network requests and application state. The most common live collection is a Stream; other live collections can be found in modules inside this crate.

Hydro uses a unique compilation approach where you define deployment logic as Rust code alongside your distributed system implementation. For more details on this API, see the Hydro docs and the deploy module.

Re-exports

pub use runtime_context::RUNTIME_CONTEXT;dfir_context

pub use boundedness::Bounded;

pub use boundedness::Unbounded;

pub use stream::NoOrder;

pub use stream::Stream;

pub use stream::TotalOrder;

pub use keyed_stream::KeyedStream;

pub use singleton::Singleton;

pub use optional::Optional;

pub use location::cluster::CLUSTER_SELF_ID;

pub use location::Atomic;

pub use location::Cluster;

pub use location::External;

pub use location::Location;

pub use location::MemberId;

pub use location::NetworkHint;

pub use location::Process;

pub use location::Tick;

pub use builder::FlowBuilder;

pub use unsafety::*;

Modules

backtrace

boundedness

builder

cycle

deploybuild

deploy_runtime

graph

Graph visualization utilities for Hydro IR

graph_utilviz

ir

keyed_singleton

keyed_stream

location

optional

rewrites

runtime_contextdfir_context

singleton

stream

test_utildeploy

unsafety

Macros

nondet

Fulfills a non-determinism guard parameter by declaring a reason why the non-determinism is tolerated or providing other non-determinism guards that forward the inner non-determinism.

q

Creates a quoted expression for Hydro.