hydro_lang::compile::builder

Struct FlowBuilder

pub struct FlowBuilder<'a> { /* private fields */ }

Implementations

impl<'a> FlowBuilder<'a>

pub fn new() -> Self

Creates a new FlowBuilder to construct a Hydro program, using the Cargo package name as the program name.

pub fn with_name(name: impl Into<String>) -> Self

Creates a new FlowBuilder to construct a Hydro program, with the given program name.

pub fn process<P>(&mut self) -> Process<'a, P>

pub fn cluster<C>(&mut self) -> Cluster<'a, C>

pub fn external<E>(&mut self) -> External<'a, E>

impl<'a> FlowBuilder<'a>

pub fn finalize(self) -> BuiltFlow<'a>

Available on crate feature build only.

pub fn with_default_optimize<D: Deploy<'a>>(self) -> DeployFlow<'a, D>

Available on crate feature build only.

pub fn optimize_with(self, f: impl FnOnce(&mut [HydroRoot])) -> BuiltFlow<'a>

Available on crate feature build only.

pub fn with_process<P, D: Deploy<'a>>( self, process: &Process<'_, P>, spec: impl IntoProcessSpec<'a, D>, ) -> DeployFlow<'a, D>

Available on crate feature build only.

pub fn with_remaining_processes<D: Deploy<'a>, S: IntoProcessSpec<'a, D> + 'a>( self, spec: impl Fn() -> S, ) -> DeployFlow<'a, D>

Available on crate feature build only.

pub fn with_external<P, D: Deploy<'a>>( self, process: &External<'_, P>, spec: impl ExternalSpec<'a, D>, ) -> DeployFlow<'a, D>

Available on crate feature build only.

pub fn with_remaining_externals<D: Deploy<'a>, S: ExternalSpec<'a, D> + 'a>( self, spec: impl Fn() -> S, ) -> DeployFlow<'a, D>

Available on crate feature build only.

pub fn with_cluster<C, D: Deploy<'a>>( self, cluster: &Cluster<'_, C>, spec: impl ClusterSpec<'a, D>, ) -> DeployFlow<'a, D>

Available on crate feature build only.

pub fn with_remaining_clusters<D: Deploy<'a>, S: ClusterSpec<'a, D> + 'a>( self, spec: impl Fn() -> S, ) -> DeployFlow<'a, D>

Available on crate feature build only.

pub fn compile<D: Deploy<'a>>(self) -> CompiledFlow<'a>

Available on crate feature build only.

pub fn deploy<D: Deploy<'a>>( self, env: &mut D::InstantiateEnv, ) -> DeployResult<'a, D>

Available on crate feature build only.

pub fn sim(self) -> SimFlow<'a>

Available on crate features build and sim only.

Creates a simulation for this builder, which can be used to run deterministic simulations of the Hydro program.

pub fn from_built<'b>(built: &BuiltFlow<'_>) -> FlowBuilder<'b>

Available on crate feature build only.