hydro_lang::compile::deploy_provider

Trait Node

pub trait Node {
    type Port: Clone;
    type Meta: Default;
    type InstantiateEnv;

    // Required methods
    fn next_port(&self) -> Self::Port;
    fn update_meta(&self, meta: &Self::Meta);
    fn instantiate(
        &self,
        env: &mut Self::InstantiateEnv,
        meta: &mut Self::Meta,
        graph: DfirGraph,
        extra_stmts: &[Stmt],
        sidecars: &[Expr],
    );
}

Available on crate feature build only.

Required Associated Types

type Port: Clone

A logical communication endpoint for this node.

Implementors are free to choose the concrete representation (for example, a handle or identifier), but it must be Clone so that a single logical port can be duplicated and passed to multiple consumers. New ports are allocated via Self::next_port.

type Meta: Default

type InstantiateEnv

Required Methods

fn next_port(&self) -> Self::Port

Allocates and returns a new port.

fn update_meta(&self, meta: &Self::Meta)

fn instantiate( &self, env: &mut Self::InstantiateEnv, meta: &mut Self::Meta, graph: DfirGraph, extra_stmts: &[Stmt], sidecars: &[Expr], )