hydro_lang::compile

Module ir

Modules

backtrace

Platform-independent interface for collecting backtraces, used in the Hydro IR to trace the origin of each node.

Structs

DebugExpr

Wrapper that displays only the tokens of a parsed expr.

DebugInstantiateFinalized

DebugType

Debug displays the type’s tokens.

HydroIrMetadata

HydroIrOpMetadata

Metadata that is specific to the operator itself, rather than its outputs. This is available on both inner nodes and roots.

QMacroSimplifier

AST visitor that simplifies q! macro expansions

TeeNode

Enums

BuildersOrCallback

DebugInstantiate

HydroNode

An intermediate node in a Hydro graph, which consumes data from upstream nodes and emits data to downstream nodes.

HydroRoot

An root in a Hydro graph, which is an pipeline that doesn’t emit any downstream values. Traversals over the dataflow graph and generating DFIR IR start from roots.

HydroSource

A source in a Hydro graph, where data enters the graph.

Functions

dbg_dedup_tee

deep_clone

emit

transform_bottom_up

traverse_dfir

Type Aliases

SeenTeeLocations

SeenTees