hydro_lang::compile::ir

Enum HydroNode

pub enum HydroNode {

    Placeholder,
    Source {
        source: HydroSource,
        metadata: HydroIrMetadata,
    },
    CycleSource {
        ident: Ident,
        metadata: HydroIrMetadata,
    },
    Tee {
        inner: TeeNode,
        metadata: HydroIrMetadata,
    },
    Persist {
        inner: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Unpersist {
        inner: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Delta {
        inner: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Chain {
        first: Box<HydroNode>,
        second: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    ChainFirst {
        first: Box<HydroNode>,
        second: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    CrossProduct {
        left: Box<HydroNode>,
        right: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    CrossSingleton {
        left: Box<HydroNode>,
        right: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Join {
        left: Box<HydroNode>,
        right: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Difference {
        pos: Box<HydroNode>,
        neg: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    AntiJoin {
        pos: Box<HydroNode>,
        neg: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    ResolveFutures {
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    ResolveFuturesOrdered {
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Map {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    FlatMap {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Filter {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    FilterMap {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    DeferTick {
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Enumerate {
        is_static: bool,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Inspect {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Unique {
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Sort {
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Fold {
        init: DebugExpr,
        acc: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Scan {
        init: DebugExpr,
        acc: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    FoldKeyed {
        init: DebugExpr,
        acc: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Reduce {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    ReduceKeyed {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    ReduceKeyedWatermark {
        f: DebugExpr,
        input: Box<HydroNode>,
        watermark: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    Network {
        serialize_fn: Option<DebugExpr>,
        instantiate_fn: DebugInstantiate,
        deserialize_fn: Option<DebugExpr>,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },
    ExternalInput {
        from_external_id: usize,
        from_key: usize,
        from_many: bool,
        codec_type: DebugType,
        port_hint: NetworkHint,
        instantiate_fn: DebugInstantiate,
        deserialize_fn: Option<DebugExpr>,
        metadata: HydroIrMetadata,
    },
    Counter {
        tag: String,
        duration: DebugExpr,
        prefix: String,
        input: Box<HydroNode>,
        metadata: HydroIrMetadata,
    },

}

An intermediate node in a Hydro graph, which consumes data from upstream nodes and emits data to downstream nodes.

Variants

Placeholder

Source

Fields

source: HydroSource

metadata: HydroIrMetadata

CycleSource

Fields

ident: Ident

metadata: HydroIrMetadata

Tee

Fields

inner: TeeNode

metadata: HydroIrMetadata

Persist

Fields

inner: Box<HydroNode>

metadata: HydroIrMetadata

Unpersist

Fields

inner: Box<HydroNode>

metadata: HydroIrMetadata

Delta

Fields

inner: Box<HydroNode>

metadata: HydroIrMetadata

Chain

Fields

first: Box<HydroNode>

second: Box<HydroNode>

metadata: HydroIrMetadata

ChainFirst

Fields

first: Box<HydroNode>

second: Box<HydroNode>

metadata: HydroIrMetadata

CrossProduct

Fields

left: Box<HydroNode>

right: Box<HydroNode>

metadata: HydroIrMetadata

CrossSingleton

Fields

left: Box<HydroNode>

right: Box<HydroNode>

metadata: HydroIrMetadata

Join

Fields

left: Box<HydroNode>

right: Box<HydroNode>

metadata: HydroIrMetadata

Difference

Fields

pos: Box<HydroNode>

neg: Box<HydroNode>

metadata: HydroIrMetadata

AntiJoin

Fields

pos: Box<HydroNode>

neg: Box<HydroNode>

metadata: HydroIrMetadata

ResolveFutures

Fields

input: Box<HydroNode>

metadata: HydroIrMetadata

ResolveFuturesOrdered

Fields

input: Box<HydroNode>

metadata: HydroIrMetadata

Map

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroIrMetadata

FlatMap

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroIrMetadata

Filter

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroIrMetadata

FilterMap

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroIrMetadata

DeferTick

Fields

input: Box<HydroNode>

metadata: HydroIrMetadata

Enumerate

Fields

is_static: bool

input: Box<HydroNode>

metadata: HydroIrMetadata

Inspect

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroIrMetadata

Unique

Fields

input: Box<HydroNode>

metadata: HydroIrMetadata

Sort

Fields

input: Box<HydroNode>

metadata: HydroIrMetadata

Fold

Fields

init: DebugExpr

acc: DebugExpr

input: Box<HydroNode>

metadata: HydroIrMetadata

Scan

Fields

init: DebugExpr

acc: DebugExpr

input: Box<HydroNode>

metadata: HydroIrMetadata

FoldKeyed

Fields

init: DebugExpr

acc: DebugExpr

input: Box<HydroNode>

metadata: HydroIrMetadata

Reduce

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroIrMetadata

ReduceKeyed

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroIrMetadata

ReduceKeyedWatermark

Fields

f: DebugExpr

input: Box<HydroNode>

watermark: Box<HydroNode>

metadata: HydroIrMetadata

Network

Fields

serialize_fn: Option<DebugExpr>

instantiate_fn: DebugInstantiate

deserialize_fn: Option<DebugExpr>

input: Box<HydroNode>

metadata: HydroIrMetadata

ExternalInput

Fields

from_external_id: usize

from_key: usize

from_many: bool

codec_type: DebugType

port_hint: NetworkHint

instantiate_fn: DebugInstantiate

deserialize_fn: Option<DebugExpr>

metadata: HydroIrMetadata

Counter

Fields

tag: String

duration: DebugExpr

prefix: String

input: Box<HydroNode>

metadata: HydroIrMetadata

Implementations

impl HydroNode

pub fn transform_bottom_up( &mut self, transform: &mut impl FnMut(&mut HydroNode), seen_tees: &mut SeenTees, check_well_formed: bool, )

pub fn transform_children( &mut self, transform: impl FnMut(&mut HydroNode, &mut SeenTees), seen_tees: &mut SeenTees, )

pub fn deep_clone(&self, seen_tees: &mut SeenTees) -> HydroNode

pub fn emit_core( &mut self, builders_or_callback: &mut BuildersOrCallback<'_, impl FnMut(&mut HydroRoot, &mut usize), impl FnMut(&mut HydroNode, &mut usize)>, built_tees: &mut HashMap<*const RefCell<HydroNode>, (Ident, usize)>, next_stmt_id: &mut usize, ) -> (Ident, usize)

pub fn visit_debug_expr(&mut self, transform: impl FnMut(&mut DebugExpr))

pub fn op_metadata(&self) -> &HydroIrOpMetadata

pub fn metadata(&self) -> &HydroIrMetadata

pub fn op_metadata_mut(&mut self) -> &mut HydroIrOpMetadata

pub fn metadata_mut(&mut self) -> &mut HydroIrMetadata

pub fn input_metadata(&self) -> Vec<&HydroIrMetadata>

pub fn print_root(&self) -> String

impl HydroNode

pub fn build_graph_structure( &self, structure: &mut HydroGraphStructure, seen_tees: &mut HashMap<*const RefCell<HydroNode>, usize>, config: &HydroWriteConfig, ) -> usize

Available on crate feature viz only.

Build the graph structure recursively for this node.