hydro_lang

Module deploy

Available on crate feature deploy only.

Infrastructure for deploying Hydro programs to the cloud using hydro_deploy.

Re-exports

pub use crate::compile::init_test;

pub use deploy_graph::*;

pub use deploy_graph_containerized::*;docker_deploy

pub use deploy_graph_containerized_ecs::*;ecs_deploy

Modules

deploy_graph

Deployment backend for Hydro that uses hydro_deploy to provision and launch services.

deploy_graph_containerizeddocker_deploy

Deployment backend for Hydro that uses Docker to provision and launch services.

deploy_graph_containerized_ecsecs_deploy

Deployment backend for Hydro that can generate manifests that can be consumed by CDK to deploy cloud formation stacks to aws.

deploy_runtime_containerizeddocker_runtime

deploy_runtime_containerized_ecsecs_runtime

maelstrommaelstrom or maelstrom_runtime

Deployment backend for running correctness tests against Jepsen Maelstrom (https://github.com/jepsen-io/maelstrom)