hydro_lang::deploy

Module deploy_graph_containerized

Available on crate features deploy and docker_deploy only.

Deployment backend for Hydro that uses Docker to provision and launch services.

Structs

DockerDeploy

For deploying to a local docker instance

DockerDeployCluster

Represents a logical cluster, which can be a variable amount of individual containers.

DockerDeployClusterSpec

Represents a Cluster running across count docker containers.

DockerDeployExternal

Represents an external process, outside the control of this deployment but still with some communication into this deployment.

DockerDeployExternalSpec

Represents an external process outside of the management of hydro deploy.

DockerDeployProcess

Represents a process running in a docker container

DockerDeployProcessSpec

Represents a Process running in a docker container

DockerNetwork

represents a docker network