hydro_lang::deploy

Module deploy_graph_containerized_ecs

Available on crate features deploy and ecs_deploy only.

Deployment backend for Hydro that can generate manifests that can be consumed by CDK to deploy cloud formation stacks to aws.

Structs

BuildConfig

Build configuration for a Hydro binary

ClusterManifest

Manifest entry for a cluster (multiple instances of the same service)

EcsDeploy

Represents an aws ecs deployment.

EcsDeployCluster

Represents a logical cluster, which can be a variable amount of individual containers.

EcsDeployClusterSpec

Represents a Cluster running across count ecs tasks.

EcsDeployExternal

Represents an external process, outside the control of this deployment but still with some communication into this deployment.

EcsDeployExternalSpec

Represents an external process outside of the management of hydro deploy.

EcsDeployProcess

Represents a process running in an ecs deployment

EcsDeployProcessSpec

Represents a Process running in an ecs deployment

HydroManifest

Manifest for CDK deployment - describes all processes, clusters, and their configuration

PortInfo

Information about an exposed port

ProcessManifest

Manifest entry for a single process