hydro_lang::graph

Module render

Available on crate feature viz only.

Re-exports

pub use super::graphviz::HydroDot;

pub use super::graphviz::escape_dot;

pub use super::mermaid::HydroMermaid;

pub use super::mermaid::escape_mermaid;

pub use super::reactflow::HydroReactFlow;

Structs

HydroGraphStructure

Graph structure tracker for Hydro IR rendering.

HydroWriteConfig

Configuration for graph writing.

IndentedGraphWriter

Base struct for text-based graph writers that use indentation. Contains common fields shared by DOT and Mermaid writers.

Enums

HydroEdgeType

Types of edges in Hydro IR.

HydroNodeType

Types of nodes in Hydro IR for styling purposes.

NodeLabel

Label for a graph node - can be either a static string or contain expressions.

Traits

HydroGraphWrite

Trait for writing textual representations of Hydro IR graphs, i.e. mermaid or dot graphs.

Functions

extract_op_name

Function to extract an op_name from a print_root() result for use in labels.

extract_short_label

Extract a short, readable label from the full token stream label using print_root() style naming

render_hydro_ir_dot

render_hydro_ir_mermaid

render_hydro_ir_reactflow

write_hydro_ir_dot

write_hydro_ir_mermaid

write_hydro_ir_reactflow

Type Aliases

GraphWriteError

Common error type used by all graph writers.