hydro_lang::ir

Enum HydroNode

pub enum HydroNode {

    Placeholder,
    Source {
        source: HydroSource,
        location_kind: LocationId,
        metadata: HydroNodeMetadata,
    },
    CycleSource {
        ident: Ident,
        location_kind: LocationId,
        metadata: HydroNodeMetadata,
    },
    Tee {
        inner: TeeNode,
        metadata: HydroNodeMetadata,
    },
    Persist {
        inner: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Unpersist {
        inner: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Delta {
        inner: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Chain {
        first: Box<HydroNode>,
        second: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    CrossProduct {
        left: Box<HydroNode>,
        right: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    CrossSingleton {
        left: Box<HydroNode>,
        right: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Join {
        left: Box<HydroNode>,
        right: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Difference {
        pos: Box<HydroNode>,
        neg: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    AntiJoin {
        pos: Box<HydroNode>,
        neg: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Map {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    FlatMap {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Filter {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    FilterMap {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    DeferTick {
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Enumerate {
        is_static: bool,
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Inspect {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Unique {
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Sort {
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Fold {
        init: DebugExpr,
        acc: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    FoldKeyed {
        init: DebugExpr,
        acc: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Reduce {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    ReduceKeyed {
        f: DebugExpr,
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },
    Network {
        from_key: Option<usize>,
        to_location: LocationId,
        to_key: Option<usize>,
        serialize_fn: Option<DebugExpr>,
        instantiate_fn: DebugInstantiate,
        deserialize_fn: Option<DebugExpr>,
        input: Box<HydroNode>,
        metadata: HydroNodeMetadata,
    },

}

An intermediate node in a Hydro graph, which consumes data from upstream nodes and emits data to downstream nodes.

Variants

Placeholder

Source

Fields

source: HydroSource

location_kind: LocationId

metadata: HydroNodeMetadata

CycleSource

Fields

ident: Ident

location_kind: LocationId

metadata: HydroNodeMetadata

Tee

Fields

inner: TeeNode

metadata: HydroNodeMetadata

Persist

Fields

inner: Box<HydroNode>

metadata: HydroNodeMetadata

Unpersist

Fields

inner: Box<HydroNode>

metadata: HydroNodeMetadata

Delta

Fields

inner: Box<HydroNode>

metadata: HydroNodeMetadata

Chain

Fields

first: Box<HydroNode>

second: Box<HydroNode>

metadata: HydroNodeMetadata

CrossProduct

Fields

left: Box<HydroNode>

right: Box<HydroNode>

metadata: HydroNodeMetadata

CrossSingleton

Fields

left: Box<HydroNode>

right: Box<HydroNode>

metadata: HydroNodeMetadata

Join

Fields

left: Box<HydroNode>

right: Box<HydroNode>

metadata: HydroNodeMetadata

Difference

Fields

pos: Box<HydroNode>

neg: Box<HydroNode>

metadata: HydroNodeMetadata

AntiJoin

Fields

pos: Box<HydroNode>

neg: Box<HydroNode>

metadata: HydroNodeMetadata

Map

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroNodeMetadata

FlatMap

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroNodeMetadata

Filter

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroNodeMetadata

FilterMap

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroNodeMetadata

DeferTick

Fields

input: Box<HydroNode>

metadata: HydroNodeMetadata

Enumerate

Fields

is_static: bool

input: Box<HydroNode>

metadata: HydroNodeMetadata

Inspect

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroNodeMetadata

Unique

Fields

input: Box<HydroNode>

metadata: HydroNodeMetadata

Sort

Fields

input: Box<HydroNode>

metadata: HydroNodeMetadata

Fold

Fields

init: DebugExpr

acc: DebugExpr

input: Box<HydroNode>

metadata: HydroNodeMetadata

FoldKeyed

Fields

init: DebugExpr

acc: DebugExpr

input: Box<HydroNode>

metadata: HydroNodeMetadata

Reduce

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroNodeMetadata

ReduceKeyed

Fields

f: DebugExpr

input: Box<HydroNode>

metadata: HydroNodeMetadata

Network

Fields

from_key: Option<usize>

to_location: LocationId

to_key: Option<usize>

serialize_fn: Option<DebugExpr>

instantiate_fn: DebugInstantiate

deserialize_fn: Option<DebugExpr>

input: Box<HydroNode>

metadata: HydroNodeMetadata

Implementations

impl<'a> HydroNode

pub fn compile_network<D: Deploy<'a>>( &mut self, compile_env: &D::CompileEnv, seen_tees: &mut SeenTees, seen_tee_locations: &mut SeenTeeLocations, nodes: &HashMap<usize, D::Process>, clusters: &HashMap<usize, D::Cluster>, externals: &HashMap<usize, D::ExternalProcess>, )

pub fn connect_network(&mut self, seen_tees: &mut SeenTees)

pub fn transform_bottom_up( &mut self, transform: &mut impl FnMut(&mut HydroNode), seen_tees: &mut SeenTees, )

pub fn transform_children( &mut self, transform: impl FnMut(&mut HydroNode, &mut SeenTees), seen_tees: &mut SeenTees, )

pub fn deep_clone(&self, seen_tees: &mut SeenTees) -> HydroNode

pub fn emit_core( &mut self, builders_or_callback: &mut BuildersOrCallback<'_, impl FnMut(&mut HydroLeaf, usize), impl FnMut(&mut HydroNode, usize)>, built_tees: &mut HashMap<*const RefCell<HydroNode>, (Ident, usize)>, next_stmt_id: &mut usize, ) -> (Ident, usize)

pub fn metadata(&self) -> &HydroNodeMetadata

pub fn print_root(&self) -> String