hydro_lang::live_collections

Module keyed_singleton

Definitions for the KeyedSingleton live collection.

Structs

BoundedValue

A variation of boundedness specific to KeyedSingleton, which indicates that once a key appears, its value is bounded and will never change. If the KeyBound is Bounded, then the entire set of entries is bounded, but if it is Unbounded, then new entries may appear asynchronously.

KeyedSingleton

Mapping from keys of type K to values of type V.

Traits

KeyedSingletonBound

A marker trait indicating which components of a KeyedSingleton may change.