hydro_lang::live_collections

Module keyed_stream

Definitions for the KeyedStream live collection.

Modules

networking

Networking APIs for KeyedStream.

Structs

KeyedStream

Streaming elements of type V grouped by a key of type K.

Enums

Generate

The output of a Hydro generator created with KeyedStream::generator, which can yield elements and control the processing of future elements.