hydro_lang::live_collections

Module optional

Definitions for the Optional live collection.

Structs

Optional

A nullable Rust value that can asynchronously change over time.

Enums

InitNone

Marks that the Optional is null only initially: once it becomes non-null it will remain non-null forever, although the non-null value may still change arbitrarily over time.

Traits

IsInitNone

Marker trait that is implemented for OptionalBound types that are null only initially: once the optional becomes non-null it remains non-null (or it is Bounded).

OptionalBound

A marker trait indicating the boundedness of an Optional.