hydro_lang::live_collections::sliced

Trait Unslicable

pub trait Unslicable {
    type Unsliced;

    // Required method
    fn unslice(self) -> Self::Unsliced;
}

A trait for live collections which can be yielded out of a slice back into their original form.

Required Associated Types

type Unsliced

The unsliced version of this live collection.

Required Methods

fn unslice(self) -> Self::Unsliced

Unslices a sliced live collection back into its original form.