hydro_lang::live_collections

Module stream

Definitions for the Stream live collection.

Modules

networking

Networking APIs for Stream.

Structs

Stream

Streaming sequence of elements with type Type.

Enums

AtLeastOnce

Marks the stream as having non-deterministic message cardinality, which means that duplicates may occur, but messages will not be dropped.

ExactlyOnce

Marks the stream as having deterministic message cardinality, with no possibility of duplicates.

NoOrder

Marks the stream as having no order, which means that the order of elements may be affected by non-determinism.

TotalOrder

Marks the stream as being totally ordered, which means that there are no sources of non-determinism (other than intentional ones) that will affect the order of elements.

Traits

MinOrder

Helper trait for determining the weakest of two orderings.

MinRetries

Helper trait for determining the weakest of two retry guarantees.

Ordering

A trait implemented by valid ordering markers (TotalOrder and NoOrder).

Retries

A trait implemented by valid retries markers (ExactlyOnce and AtLeastOnce).

WeakerOrderingThan

Marker trait for an Ordering that is available when Self is a weaker guarantee than Other, which means that a stream with Other guarantees can be safely converted to have Self guarantees instead.

WeakerRetryThan

Marker trait for a Retries that is available when Self is a weaker guarantee than Other, which means that a stream with Other guarantees can be safely converted to have Self guarantees instead.