hydro_lang::location

Trait Location

pub trait Location<'a>: DynLocation {
    type Root: Location<'a>;
    type DropConsistency: Location<'a, DropConsistency = Self::DropConsistency>;

    // Required methods
    fn root(&self) -> Self::Root;
    fn drop_consistency(&self) -> Self::DropConsistency;
    fn consistency() -> Option<ClusterConsistency>;

    // Provided methods
    fn with_consistency_of<L2: Location<'a, DropConsistency = Self::DropConsistency>>(
        &self,
    ) -> L2 { ... }
    fn try_tick(&self) -> Option<Tick<Self>> { ... }
    fn id(&self) -> LocationId { ... }
    fn tick(&self) -> Tick<Self> { ... }
    fn spin(&self) -> Stream<(), Self, Unbounded, TotalOrder, ExactlyOnce>
       where Self: TopLevel<'a> + Sized { ... }
    fn source_stream<T, E>(
        &self,
        e: impl QuotedWithContext<'a, E, Self>,
    ) -> Stream<T, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>
       where E: FuturesStream<Item = T> + Unpin,
             Self: TopLevel<'a> + Sized { ... }
    fn source_iter<T, E>(
        &self,
        e: impl QuotedWithContext<'a, E, Self>,
    ) -> Stream<T, Self::DropConsistency, Bounded, TotalOrder, ExactlyOnce>
       where E: IntoIterator<Item = T>,
             Self: Sized { ... }
    fn source_cluster_members<C: 'a>(
        &self,
        cluster: &Cluster<'a, C>,
        nondet_start: NonDet,
    ) -> KeyedStream<MemberId<C>, MembershipEvent, Self::DropConsistency, Unbounded>
       where Self: TopLevel<'a> + Sized { ... }
    fn source_cluster_membership_stream<C: 'a>(
        &self,
        cluster: &Cluster<'a, C>,
        _nondet_start: NonDet,
    ) -> KeyedStream<MemberId<C>, MembershipEvent, Self::DropConsistency, Unbounded>
       where Self: TopLevel<'a> + Sized { ... }
    fn source_external_bytes<L>(
        &self,
        from: &External<'_, L>,
    ) -> (ExternalBytesPort, Stream<BytesMut, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>)
       where Self: TopLevel<'a> + Sized { ... }
    fn source_external_bincode<L, T, O: Ordering, R: Retries>(
        &self,
        from: &External<'_, L>,
    ) -> (ExternalBincodeSink<T, NotMany, O, R>, Stream<T, Self::DropConsistency, Unbounded, O, R>)
       where Self: TopLevel<'a> + Sized,
             T: Serialize + DeserializeOwned { ... }
    fn sim_input<T, O: Ordering, R: Retries>(
        &self,
    ) -> (SimSender<T, O, R>, Stream<T, Self::DropConsistency, Unbounded, O, R>)
       where Self: TopLevel<'a> + Sized,
             T: Serialize + DeserializeOwned { ... }
    fn embedded_input<T>(
        &self,
        name: impl Into<String>,
    ) -> Stream<T, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>
       where Self: TopLevel<'a> + Sized { ... }
    fn embedded_singleton_input<T>(
        &self,
        name: impl Into<String>,
    ) -> Singleton<T, Self::DropConsistency, Bounded>
       where Self: TopLevel<'a> + Sized { ... }
    fn bind_single_client<L, T, Codec: Encoder<T> + Decoder>(
        &self,
        from: &External<'_, L>,
        port_hint: NetworkHint,
    ) -> (ExternalBytesPort<NotMany>, Stream<<Codec as Decoder>::Item, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, Stream<T, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>>)
       where Self: TopLevel<'a> + Sized { ... }
    fn bind_single_client_bincode<L, InT: DeserializeOwned, OutT: Serialize>(
        &self,
        from: &External<'_, L>,
    ) -> (ExternalBincodeBidi<InT, OutT, NotMany>, Stream<InT, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, Stream<OutT, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>>)
       where Self: TopLevel<'a> + Sized { ... }
    fn bidi_external_many_bytes<L, T, Codec: Encoder<T> + Decoder>(
        &self,
        from: &External<'_, L>,
        port_hint: NetworkHint,
    ) -> (ExternalBytesPort<Many>, KeyedStream<u64, <Codec as Decoder>::Item, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>, KeyedStream<u64, MembershipEvent, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, KeyedStream<u64, T, Self::DropConsistency, Unbounded, NoOrder, ExactlyOnce>>)
       where Self: TopLevel<'a> + Sized { ... }
    fn bidi_external_many_bincode<L, InT: DeserializeOwned, OutT: Serialize>(
        &self,
        from: &External<'_, L>,
    ) -> (ExternalBincodeBidi<InT, OutT, Many>, KeyedStream<u64, InT, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>, KeyedStream<u64, MembershipEvent, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, KeyedStream<u64, OutT, Self::DropConsistency, Unbounded, NoOrder, ExactlyOnce>>)
       where Self: TopLevel<'a> + Sized { ... }
    fn sidecar_bidi<InT: 'static, OutT: 'static, F>(
        &self,
        sidecar: impl QuotedWithContext<'a, F, Self>,
    ) -> (Stream<InT, Self, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, Stream<OutT, Self, Unbounded, NoOrder, ExactlyOnce>>)
       where Self: Sized + TopLevel<'a> { ... }
    fn singleton<T>(
        &self,
        e: impl QuotedWithContext<'a, T, Self>,
    ) -> Singleton<T, Self::DropConsistency, Bounded>
       where Self: Sized { ... }
    fn singleton_future<F>(
        &self,
        e: impl QuotedWithContext<'a, F, Self>,
    ) -> Singleton<F::Output, Self::DropConsistency, Bounded>
       where F: Future,
             Self: Sized { ... }
    fn source_interval(
        &self,
        interval: impl QuotedWithContext<'a, Duration, Self> + Copy + 'a,
    ) -> Stream<(), Self, Unbounded, TotalOrder, ExactlyOnce>
       where Self: TopLevel<'a> + Sized { ... }
    fn source_interval_delayed(
        &self,
        delay: impl QuotedWithContext<'a, Duration, Self> + Copy + 'a,
        interval: impl QuotedWithContext<'a, Duration, Self> + Copy + 'a,
    ) -> Stream<(), Self, Unbounded, TotalOrder, ExactlyOnce>
       where Self: TopLevel<'a> + Sized { ... }
    fn forward_ref<S>(&self) -> (ForwardHandle<'a, S>, S)
       where S: CycleCollection<'a, ForwardRef, Location = Self> { ... }

}

A location where data can be materialized and computation can be executed.

Hydro is a global, distributed programming model. This means that the data and computation in a Hydro program can be spread across multiple machines, data centers, and even continents. To achieve this, Hydro uses the concept of locations to keep track of where data is located and computation is executed.

Each live collection type (in crate::live_collections) has a type parameter L which will always be a type that implements the Location trait (e.g. Process and Cluster). To create distributed programs, Hydro provides a variety of APIs to allow live collections to be moved between locations via network send/receive.

See the Hydro docs for more information.

Required Associated Types

type Root: Location<'a>

The root location type for this location.

For top-level locations like Process and Cluster, this is Self. For nested locations like Tick, this is the root location that contains it.

type DropConsistency: Location<'a, DropConsistency = Self::DropConsistency>

Location type with consistency guarantees dropped for the live collection on it.

Required Methods

fn root(&self) -> Self::Root

Returns the root location for this location.

For top-level locations like Process and Cluster, this returns self. For nested locations like Tick, this returns the root location that contains it.

fn drop_consistency(&self) -> Self::DropConsistency

This location but with consistency guarantees dropped for the live collection

fn consistency() -> Option<ClusterConsistency>

Gets the runtime enum variant for the current consistency level, if this is a cluster.

Provided Methods

fn with_consistency_of<L2: Location<'a, DropConsistency = Self::DropConsistency>>( &self, ) -> L2

Updates the consistency guarantees to match that of the given location.

fn try_tick(&self) -> Option<Tick<Self>>

Attempts to create a new Tick clock domain at this location.

Returns Some(Tick) if this is a top-level location (like Process or Cluster), or None if this location is already inside a tick (nested ticks are not supported).

Prefer using Location::tick when you know the location is top-level.

fn id(&self) -> LocationId

Returns the unique identifier for this location.

fn tick(&self) -> Tick<Self>

Creates a new Tick clock domain at this location.

A tick represents a logical clock that can be used to batch streaming data into discrete time steps. This is useful for implementing iterative algorithms or for synchronizing data across multiple streams.

Example

let tick = process.tick();
let inside_tick = process
    .source_iter(q!(vec![1, 2, 3, 4]))
    .batch(&tick, nondet!(/** test */));
inside_tick.all_ticks()
// 1, 2, 3, 4

fn spin(&self) -> Stream<(), Self, Unbounded, TotalOrder, ExactlyOnce>
where Self: TopLevel<'a> + Sized,

Creates an unbounded stream that continuously emits unit values ().

This is useful for driving computations that need to run continuously, such as polling or heartbeat mechanisms.

Example

let tick = process.tick();
process.spin()
    .batch(&tick, nondet!(/** test */))
    .map(q!(|_| 42))
    .all_ticks()
// 42, 42, 42, ...

fn source_stream<T, E>( &self, e: impl QuotedWithContext<'a, E, Self>, ) -> Stream<T, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>
where E: FuturesStream<Item = T> + Unpin, Self: TopLevel<'a> + Sized,

Creates a stream from an async [FuturesStream].

This is useful for integrating with external async data sources, such as network connections or file readers.

Example

process.source_stream(q!(futures::stream::iter(vec![1, 2, 3])))
// 1, 2, 3

fn source_iter<T, E>( &self, e: impl QuotedWithContext<'a, E, Self>, ) -> Stream<T, Self::DropConsistency, Bounded, TotalOrder, ExactlyOnce>
where E: IntoIterator<Item = T>, Self: Sized,

Creates a bounded stream from an iterator.

The iterator is evaluated once at runtime, and all elements are emitted in order. This is useful for creating streams from static data or for testing.

Example

process.source_iter(q!(vec![1, 2, 3, 4]))
// 1, 2, 3, 4

fn source_cluster_members<C: 'a>( &self, cluster: &Cluster<'a, C>, nondet_start: NonDet, ) -> KeyedStream<MemberId<C>, MembershipEvent, Self::DropConsistency, Unbounded>
where Self: TopLevel<'a> + Sized,

👎Deprecated:
use .source_cluster_membership_stream(…) instead

Creates a stream of membership events for a cluster.

This stream emits MembershipEvent::Joined when a cluster member joins and MembershipEvent::Left when a cluster member leaves. The stream is keyed by the MemberId of the cluster member.

This is useful for implementing protocols that need to track cluster membership, such as broadcasting to all members or detecting failures.

Non-Determinism

This stream is non-deterministic because the timing of membership events, for example if a node leaves, the membership event may not be received if the node left before the stream was created.

Example

let p1 = flow.process::<()>();
let workers: Cluster<()> = flow.cluster::<()>();
let cluster_members = p1.source_cluster_members(&workers, nondet!(/** late joiners may miss events */));
// if there are 4 members in the cluster, we would see a join event for each
// { MemberId::<Worker>(0): [MembershipEvent::Join], MemberId::<Worker>(2): [MembershipEvent::Join], ... }

fn source_cluster_membership_stream<C: 'a>( &self, cluster: &Cluster<'a, C>, _nondet_start: NonDet, ) -> KeyedStream<MemberId<C>, MembershipEvent, Self::DropConsistency, Unbounded>
where Self: TopLevel<'a> + Sized,

Creates a stream of membership events for a cluster.

This stream emits MembershipEvent::Joined when a cluster member joins and MembershipEvent::Left when a cluster member leaves. The stream is keyed by the MemberId of the cluster member.

This is useful for implementing protocols that need to track cluster membership, such as broadcasting to all members or detecting failures.

Non-Determinism

This stream is non-deterministic because the timing of membership events, for example if a node leaves, the membership event may not be received if the node left before the stream was created.

Example

let p1 = flow.process::<()>();
let workers: Cluster<()> = flow.cluster::<()>();
let cluster_members = p1.source_cluster_membership_stream(&workers, nondet!(/** late joiners may miss events */));
// if there are 4 members in the cluster, we would see a join event for each
// { MemberId::<Worker>(0): [MembershipEvent::Join], MemberId::<Worker>(2): [MembershipEvent::Join], ... }

fn source_external_bytes<L>( &self, from: &External<'_, L>, ) -> (ExternalBytesPort, Stream<BytesMut, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>)
where Self: TopLevel<'a> + Sized,

Creates a one-way connection from an external process to receive raw bytes.

Returns a port handle for the external process to connect to, and a stream of received byte buffers.

For bidirectional communication or typed data, see Location::bind_single_client or Location::source_external_bincode.

fn source_external_bincode<L, T, O: Ordering, R: Retries>( &self, from: &External<'_, L>, ) -> (ExternalBincodeSink<T, NotMany, O, R>, Stream<T, Self::DropConsistency, Unbounded, O, R>)
where Self: TopLevel<'a> + Sized, T: Serialize + DeserializeOwned,

Creates a one-way connection from an external process to receive bincode-serialized data.

Returns a sink handle for the external process to send data to, and a stream of received values.

For bidirectional communication, see Location::bind_single_client_bincode.

fn sim_input<T, O: Ordering, R: Retries>( &self, ) -> (SimSender<T, O, R>, Stream<T, Self::DropConsistency, Unbounded, O, R>)
where Self: TopLevel<'a> + Sized, T: Serialize + DeserializeOwned,

Available on crate feature sim only.

Sets up a simulated input port on this location for testing.

Returns a handle to send messages to the location as well as a stream of received messages. This is only available when the sim feature is enabled.

fn embedded_input<T>( &self, name: impl Into<String>, ) -> Stream<T, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>
where Self: TopLevel<'a> + Sized,

Creates an external input stream for embedded deployment mode.

The name parameter specifies the name of the generated function parameter that will supply data to this stream at runtime. The generated function will accept an impl Stream<Item = T> + Unpin argument with this name.

fn embedded_singleton_input<T>( &self, name: impl Into<String>, ) -> Singleton<T, Self::DropConsistency, Bounded>
where Self: TopLevel<'a> + Sized,

Creates an embedded singleton input for embedded deployment mode.

The name parameter specifies the name of the generated function parameter that will supply data to this singleton at runtime. The generated function will accept a plain T parameter with this name.

fn bind_single_client<L, T, Codec: Encoder<T> + Decoder>( &self, from: &External<'_, L>, port_hint: NetworkHint, ) -> (ExternalBytesPort<NotMany>, Stream<<Codec as Decoder>::Item, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, Stream<T, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>>)
where Self: TopLevel<'a> + Sized,

Establishes a server on this location to receive a bidirectional connection from a single client, identified by the given External handle. Returns a port handle for the external process to connect to, a stream of incoming messages, and a handle to send outgoing messages.

Example

let node = flow.process::<()>();
let external = flow.external::<()>();
let (port, incoming, outgoing) =
    node.bind_single_client::<_, Bytes, LengthDelimitedCodec>(&external, NetworkHint::Auto);
outgoing.complete(incoming.map(q!(|data /* : Bytes */| {
    let mut resp: Vec<u8> = data.into();
    resp.push(42);
    resp.into() // : Bytes
})));

let nodes = flow // ... with_process and with_external

deployment.deploy().await.unwrap();
deployment.start().await.unwrap();

let (mut external_out, mut external_in) = nodes.connect(port).await;
external_in.send(vec![1, 2, 3].into()).await.unwrap();
assert_eq!(
    external_out.next().await.unwrap().unwrap(),
    vec![1, 2, 3, 42]
);

fn bind_single_client_bincode<L, InT: DeserializeOwned, OutT: Serialize>( &self, from: &External<'_, L>, ) -> (ExternalBincodeBidi<InT, OutT, NotMany>, Stream<InT, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, Stream<OutT, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>>)
where Self: TopLevel<'a> + Sized,

Establishes a bidirectional connection from a single external client using bincode serialization.

Returns a port handle for the external process to connect to, a stream of incoming messages, and a handle to send outgoing messages. This is a convenience wrapper around Location::bind_single_client that uses bincode for serialization.

Type Parameters

InT: The type of incoming messages (must implement DeserializeOwned)

OutT: The type of outgoing messages (must implement Serialize)

fn bidi_external_many_bytes<L, T, Codec: Encoder<T> + Decoder>( &self, from: &External<'_, L>, port_hint: NetworkHint, ) -> (ExternalBytesPort<Many>, KeyedStream<u64, <Codec as Decoder>::Item, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>, KeyedStream<u64, MembershipEvent, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, KeyedStream<u64, T, Self::DropConsistency, Unbounded, NoOrder, ExactlyOnce>>)
where Self: TopLevel<'a> + Sized,

Establishes a server on this location to receive bidirectional connections from multiple external clients using raw bytes.

Unlike Location::bind_single_client, this method supports multiple concurrent client connections. Each client is assigned a unique u64 identifier.

Returns:

A port handle for external processes to connect to

A keyed stream of incoming messages, keyed by client ID

A keyed stream of membership events (client joins/leaves), keyed by client ID

A handle to send outgoing messages, keyed by client ID

fn bidi_external_many_bincode<L, InT: DeserializeOwned, OutT: Serialize>( &self, from: &External<'_, L>, ) -> (ExternalBincodeBidi<InT, OutT, Many>, KeyedStream<u64, InT, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>, KeyedStream<u64, MembershipEvent, Self::DropConsistency, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, KeyedStream<u64, OutT, Self::DropConsistency, Unbounded, NoOrder, ExactlyOnce>>)
where Self: TopLevel<'a> + Sized,

Establishes a server on this location to receive bidirectional connections from multiple external clients using bincode serialization.

Unlike Location::bind_single_client_bincode, this method supports multiple concurrent client connections. Each client is assigned a unique u64 identifier.

Returns:

A port handle for external processes to connect to

A keyed stream of incoming messages, keyed by client ID

A keyed stream of membership events (client joins/leaves), keyed by client ID

A handle to send outgoing messages, keyed by client ID

Type Parameters

InT: The type of incoming messages (must implement DeserializeOwned)

OutT: The type of outgoing messages (must implement Serialize)

fn sidecar_bidi<InT: 'static, OutT: 'static, F>( &self, sidecar: impl QuotedWithContext<'a, F, Self>, ) -> (Stream<InT, Self, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, Stream<OutT, Self, Unbounded, NoOrder, ExactlyOnce>>)
where Self: Sized + TopLevel<'a>,

Bridges user-owned async code to the dataflow as a bidirectional sidecar.

The closure is called once at startup and must return a (Stream<InT>, Sink<OutT>) pair. The framework reads from the stream (items flowing into the dataflow) and writes to the sink (items flowing out to the sidecar). The user controls buffering, backpressure, and internal lifecycle — Hydro only sees the stream/sink interface.

This will hopefully make it easy to integrate hydro with existing frameworks, for example grpc code generated service endpoints.

Returns

A Stream<InT> carrying items from the sidecar into the dataflow.

A ForwardHandle expecting a Stream<OutT> that the user completes with items destined for the sidecar.

Example

// Sidecar that echoes whatever it receives back into the dataflow.
let (inbound, response_handle) = process.sidecar_bidi::<String, String, _>(q!(|| {
    let (to_df_tx, to_df_rx) = tokio::sync::mpsc::channel::<String>(16);
    let (from_df_tx, mut from_df_rx) = tokio::sync::mpsc::channel::<String>(16);

    // Spawn the sidecar: echoes items from the dataflow back into it.
    tokio::spawn(async move {
        while let Some(msg) = from_df_rx.recv().await {
            to_df_tx.send(msg).await.ok();
        }
    });

    // Return the framework-facing ends (concrete types, no boxing needed).
    let stream = tokio_stream::wrappers::ReceiverStream::new(to_df_rx);
    let sink = tokio_util::sync::PollSender::new(from_df_tx);
    (stream, sink)
}));

// Send "hello" into the sidecar via the response channel.
let input = process.source_stream(q!(futures::stream::iter(vec!["hello".to_string()])));
response_handle.complete(input);

// The sidecar echoes it back — assert we get "hello" out.
inbound

fn singleton<T>( &self, e: impl QuotedWithContext<'a, T, Self>, ) -> Singleton<T, Self::DropConsistency, Bounded>
where Self: Sized,

Constructs a Singleton materialized at this location with the given static value.

See also: Tick::singleton, for creating a singleton within a tick, which requires T: Clone.

Example

let singleton = process.singleton(q!(5));
// 5

fn singleton_future<F>( &self, e: impl QuotedWithContext<'a, F, Self>, ) -> Singleton<F::Output, Self::DropConsistency, Bounded>
where F: Future, Self: Sized,

Constructs a Singleton by resolving an async Future to completion.

This is a convenience method equivalent to self.singleton(future_expr).resolve_future_blocking(), which is a common pattern when initializing a singleton from an async computation.

Example

let singleton = process.singleton_future(q!(async { 42 }));
singleton.into_stream()
// 42

fn source_interval( &self, interval: impl QuotedWithContext<'a, Duration, Self> + Copy + 'a, ) -> Stream<(), Self, Unbounded, TotalOrder, ExactlyOnce>
where Self: TopLevel<'a> + Sized,

Generates a stream that emits () at a fixed interval.

The first tick completes immediately. Missed ticks will be scheduled as soon as possible.

Because this only emits (), the non-determinism of when events fire is captured by the AtLeastOnce retry semantics downstream, so no NonDet guard is required.

fn source_interval_delayed( &self, delay: impl QuotedWithContext<'a, Duration, Self> + Copy + 'a, interval: impl QuotedWithContext<'a, Duration, Self> + Copy + 'a, ) -> Stream<(), Self, Unbounded, TotalOrder, ExactlyOnce>
where Self: TopLevel<'a> + Sized,

Generates a stream that emits () at a fixed interval, after an initial delay.

Because this only emits (), the non-determinism of when events fire is captured by the AtLeastOnce retry semantics downstream, so no NonDet guard is required.

fn forward_ref<S>(&self) -> (ForwardHandle<'a, S>, S)
where S: CycleCollection<'a, ForwardRef, Location = Self>,

Creates a forward reference, allowing a stream to be used before its source is defined.

Returns a (handle, placeholder) pair. Use the placeholder in the dataflow graph, then call handle.complete(actual_stream) to wire in the real source.

This is useful for mutually-dependent dataflows or when the definition order doesn’t match the data flow direction. For feedback loops, prefer Tick::cycle instead, which automatically defers values by one tick.

Panics

Panics if the forward reference creates a synchronous cycle (i.e., the completed stream transitively depends on the placeholder without a defer_tick or network hop in between).

Example

// Create a forward reference to define a stream that will be completed later
let (complete, forward_stream) = process.forward_ref::<Stream<i32, _, _, NoOrder>>();

// Use the forward reference as input to another computation
let output: Stream<_, _, _, NoOrder> = forward_stream.map(q!(|x| x * 2));

// Complete the forward reference with the actual source
let source: Stream<_, _, Unbounded> = process.source_iter(q!([1, 2, 3])).into();
complete.complete(source);
output
// 2, 4, 6

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".