hydro_lang::optional

Struct Optional

pub struct Optional<Type, Loc, Bound> { /* private fields */ }

Implementations

impl<'a, T, L, B> Optional<T, L, B>
where L: Location<'a>,

pub fn some(singleton: Singleton<T, L, B>) -> Self

impl<'a, T, L, B> Optional<T, L, B>
where L: Location<'a>,

pub fn map<U, F>(self, f: impl IntoQuotedMut<'a, F, L>) -> Optional<U, L, B>
where F: Fn(T) -> U + 'a,

Transforms the optional value by applying a function f to it, continuously as the input is updated.

Whenever the optional is empty, the output optional is also empty.

Example

let tick = process.tick();
let optional = tick.optional_first_tick(q!(1));
optional.map(q!(|v| v + 1)).all_ticks()
// 2

pub fn flat_map_ordered<U, I, F>( self, f: impl IntoQuotedMut<'a, F, L>, ) -> Stream<U, L, B>
where I: IntoIterator<Item = U>, F: Fn(T) -> I + 'a,

pub fn flat_map_unordered<U, I, F>( self, f: impl IntoQuotedMut<'a, F, L>, ) -> Stream<U, L, B, NoOrder>
where I: IntoIterator<Item = U>, F: Fn(T) -> I + 'a,

pub fn flatten_ordered<U>(self) -> Stream<U, L, B>
where T: IntoIterator<Item = U>,

pub fn flatten_unordered<U>(self) -> Stream<U, L, B, NoOrder>
where T: IntoIterator<Item = U>,

pub fn filter<F>(self, f: impl IntoQuotedMut<'a, F, L>) -> Optional<T, L, B>
where F: Fn(&T) -> bool + 'a,

pub fn filter_map<U, F>( self, f: impl IntoQuotedMut<'a, F, L>, ) -> Optional<U, L, B>
where F: Fn(T) -> Option<U> + 'a,

pub fn union(self, other: Optional<T, L, B>) -> Optional<T, L, B>

pub fn zip<O>( self, other: impl Into<Optional<O, L, B>>, ) -> Optional<(T, O), L, B>
where O: Clone,

pub fn unwrap_or(self, other: Singleton<T, L, B>) -> Singleton<T, L, B>

pub fn into_singleton(self) -> Singleton<Option<T>, L, B>
where T: Clone,

impl<'a, T, L> Optional<T, L, Bounded>
where L: Location<'a>,

pub fn continue_if<U>( self, signal: Optional<U, L, Bounded>, ) -> Optional<T, L, Bounded>

pub fn continue_unless<U>( self, other: Optional<U, L, Bounded>, ) -> Optional<T, L, Bounded>

pub fn then<U>(self, value: Singleton<U, L, Bounded>) -> Optional<U, L, Bounded>
where Singleton<U, L, Bounded>: ZipResult<'a, Optional<(), L, Bounded>, Location = L, Out = Optional<(U, ()), L, Bounded>>,

pub fn into_stream(self) -> Stream<T, L, Bounded>

impl<'a, T, L, B> Optional<T, Atomic<L>, B>
where L: Location<'a> + NoTick,

pub unsafe fn latest_tick(self) -> Optional<T, Tick<L>, Bounded>

Returns an optional value corresponding to the latest snapshot of the optional being atomically processed. The snapshot at tick t + 1 is guaranteed to include at least all relevant data that contributed to the snapshot at tick t.

Safety

Because this picks a snapshot of a optional whose value is continuously changing, the output optional has a non-deterministic value since the snapshot can be at an arbitrary point in time.

pub fn end_atomic(self) -> Optional<T, L, B>

impl<'a, T, L, B> Optional<T, L, B>
where L: Location<'a> + NoTick + NoAtomic,

pub fn atomic(self, tick: &Tick<L>) -> Optional<T, Atomic<L>, B>

pub unsafe fn latest_tick(self, tick: &Tick<L>) -> Optional<T, Tick<L>, Bounded>

Given a tick, returns a optional value corresponding to a snapshot of the optional as of that tick. The snapshot at tick t + 1 is guaranteed to include at least all relevant data that contributed to the snapshot at tick t.

Safety

Because this picks a snapshot of a optional whose value is continuously changing, the output optional has a non-deterministic value since the snapshot can be at an arbitrary point in time.

pub unsafe fn sample_eager(self) -> Stream<T, L, Unbounded>

Eagerly samples the optional as fast as possible, returning a stream of snapshots with order corresponding to increasing prefixes of data contributing to the optional.

Safety

At runtime, the optional will be arbitrarily sampled as fast as possible, but due to non-deterministic batching and arrival of inputs, the output stream is non-deterministic.

pub unsafe fn sample_every( self, interval: impl QuotedWithContext<'a, Duration, L> + Copy + 'a, ) -> Stream<T, L, Unbounded>

Given a time interval, returns a stream corresponding to snapshots of the optional value taken at various points in time. Because the input optional may be Unbounded, there are no guarantees on what these snapshots are other than they represent the value of the optional given some prefix of the streams leading up to it.

Safety

The output stream is non-deterministic in which elements are sampled, since this is controlled by a clock.

impl<'a, T, L> Optional<T, Tick<L>, Bounded>
where L: Location<'a>,

pub fn all_ticks(self) -> Stream<T, L, Unbounded>

pub fn all_ticks_atomic(self) -> Stream<T, Atomic<L>, Unbounded>

pub fn latest(self) -> Optional<T, L, Unbounded>

pub fn latest_atomic(self) -> Optional<T, Atomic<L>, Unbounded>

pub fn defer_tick(self) -> Optional<T, Tick<L>, Bounded>

pub fn persist(self) -> Stream<T, Tick<L>, Bounded>

pub fn delta(self) -> Optional<T, Tick<L>, Bounded>