hydro_lang

Module prelude

A “prelude” for users of the hydro_lang crate.

This prelude is similar to the standard library’s prelude in that you’ll almost always want to import its entire contents, but unlike the standard library’s prelude you’ll have to do so manually:

use hydro_lang::prelude::*;

The prelude may grow over time as additional items see ubiquitous use.

Re-exports

pub use crate::compile::builder::FlowBuilder;

pub use crate::live_collections::boundedness::Bounded;

pub use crate::live_collections::boundedness::Unbounded;

pub use crate::live_collections::keyed_singleton::KeyedSingleton;

pub use crate::live_collections::keyed_stream::KeyedStream;

pub use crate::live_collections::optional::Optional;

pub use crate::live_collections::singleton::Singleton;

pub use crate::live_collections::sliced::sliced;

pub use crate::live_collections::stream::Stream;

pub use crate::location::Cluster;

pub use crate::location::External;

pub use crate::location::Location as _;

pub use crate::location::Process;

pub use crate::location::Tick;

pub use crate::networking::TCP;

pub use crate::nondet::NonDet;

pub use crate::nondet::nondet;

pub use crate::properties::ManualProof;

Macros

q

Creates a quoted expression for Hydro.