hydro_lang

Module properties

Types for reasoning about algebraic properties for Rust closures.

Structs

AggFuncAlgebra

Algebraic properties for an aggregation function of type (T, &mut A) -> ().

ManualProof

A hand-written human proof of the correctness property.

Enums

NotProved

Marks that the property is not proved.

Proved

Marks that the property is proven.

Traits

CommutativeProof

A trait for proof mechanisms that can validate commutativity.

IdempotentProof

A trait for proof mechanisms that can validate idempotence.

ValidCommutativityFor

Marker trait identifying that the commutativity property is valid for the given stream ordering.

ValidIdempotenceFor

Marker trait identifying that the idempotence property is valid for the given stream ordering.