hydro_lang::properties

Trait ApplyMonotoneStream

pub trait ApplyMonotoneStream<P, B2: SingletonBound>: Sealed<P, B2> { }

Marker trait identifying the boundedness of a singleton given a monotonicity property of an aggregation on a stream.