hydro_lang

Module stream

Structs

NoOrder

Marks the stream as having no order, which means that the order of elements may be affected by non-determinism.

Stream

An ordered sequence stream of elements of type T.

TotalOrder

Marks the stream as being totally ordered, which means that there are no sources of non-determinism (other than intentional ones) that will affect the order of elements.

Traits

MinOrder

Helper trait for determining the weakest of two orderings.

Functions

deserialize_bincode_with_type

serialize_bincode_with_type