hydro_lang

Module test_util

Available on crate feature deploy only.

Various utilities for testing short Hydro programs, especially in doctests.

Functions

assert_panics_with_message

Asserts that running the given closure results in a panic with a message containing msg.

multi_location_test

Sets up a test with multiple processes / clusters declared in the test logic (thunk). The test logic must return a single streaming output, which can then be read in check (an async closure) to perform assertions.

stream_transform_test

Sets up a test declared in thunk that executes on a single Process, returning a streaming output that can be read in check (an async closure) to perform assertions.