Crate hydro_lang

Hydro is a high-level distributed programming framework for Rust. Hydro can help you quickly write scalable distributed services that are correct by construction. Much like Rust helps with memory safety, Hydro helps with distributed safety.

The core Hydro API involves live collections, which represent asynchronously updated sources of data such as incoming network requests and application state. The most common live collection is live_collections::stream::Stream; other live collections can be found in live_collections.

Hydro uses a unique compilation approach where you define deployment logic as Rust code alongside your distributed system implementation. For more details on this API, see the Hydro docs and the deploy module.

Modules

compile

Hydro compilation: the Hydro IR, Hydro to DFIR translation, and traits for deployment targets.

deploydeploy

Infrastructure for deploying Hydro programs to the cloud using hydro_deploy.

forward_handle

Mechanisms for introducing forward references and cycles in Hydro.

live_collections

Definitions for live collections, which offer the core APIs for writing distributed applications.

location

Type definitions for distributed locations, which specify where pieces of a Hydro program will be executed.

networking

Types for configuring network channels with serialization formats, transports, etc.

nondet

Defines the NonDet type and nondet! macro for tracking non-determinism.

prelude

A “prelude” for users of the hydro_lang crate.

properties

Types for reasoning about algebraic properties for Rust closures.

runtime_contextdfir_context

Interfaces to access DFIR runtime metadata from Hydro programs.

simsim

Deterministic simulation testing support for Hydro programs.

telemetry

Telemetry

test_utildeploy

Various utilities for testing short Hydro programs, especially in doctests.

vizviz

Graph visualization utilities for Hydro IR

Macros

__nondet__

Fulfills a non-determinism guard parameter by declaring a reason why the non-determinism is tolerated or providing other non-determinism guards that forward the inner non-determinism.

__sliced__

Transforms a live collection with a computation relying on a slice of another live collection. This is useful for reading a snapshot of an asynchronously updated collection while processing another collection, such as joining a stream with the latest values from a singleton.

setup

A macro to set up a Hydro crate.