hydro_lang::compile::built

Struct BuiltFlow

pub struct BuiltFlow<'a> { /* private fields */ }

Available on crate feature build only.

Implementations

impl<'a> BuiltFlow<'a>

pub fn ir(&self) -> &[HydroRoot]

Returns all HydroRoots in the IR.

pub fn location_names(&self) -> &SecondaryMap<LocationKey, String>

Returns all raw location ID -> location name mappings.

pub fn graph_api(&self) -> GraphApi<'_>

Available on stageleft_runtime and crate feature viz only.

Get a GraphApi instance for this built flow

pub fn mermaid_string( &self, show_metadata: bool, show_location_groups: bool, use_short_labels: bool, ) -> String

Available on crate feature viz only.

pub fn dot_string( &self, show_metadata: bool, show_location_groups: bool, use_short_labels: bool, ) -> String

Available on crate feature viz only.

pub fn hydroscope_string( &self, show_metadata: bool, show_location_groups: bool, use_short_labels: bool, ) -> String

Available on crate feature viz only.

pub fn mermaid_to_file( &self, filename: &str, show_metadata: bool, show_location_groups: bool, use_short_labels: bool, ) -> Result<(), Box<dyn Error>>

Available on crate feature viz only.

pub fn dot_to_file( &self, filename: &str, show_metadata: bool, show_location_groups: bool, use_short_labels: bool, ) -> Result<(), Box<dyn Error>>

Available on crate feature viz only.

pub fn hydroscope_to_file( &self, filename: &str, show_metadata: bool, show_location_groups: bool, use_short_labels: bool, ) -> Result<(), Box<dyn Error>>

Available on crate feature viz only.

pub fn mermaid_to_browser( &self, show_metadata: bool, show_location_groups: bool, use_short_labels: bool, message_handler: Option<&dyn Fn(&str)>, ) -> Result<(), Box<dyn Error>>

Available on crate feature viz only.

pub fn dot_to_browser( &self, show_metadata: bool, show_location_groups: bool, use_short_labels: bool, message_handler: Option<&dyn Fn(&str)>, ) -> Result<(), Box<dyn Error>>

Available on crate feature viz only.

pub fn hydroscope_to_browser( &self, show_metadata: bool, show_location_groups: bool, use_short_labels: bool, message_handler: Option<&dyn Fn(&str)>, ) -> Result<(), Box<dyn Error>>

Available on crate feature viz only.

pub fn optimize_with(self, f: impl FnOnce(&mut [HydroRoot])) -> Self

pub fn with_default_optimize<D: Deploy<'a>>(self) -> DeployFlow<'a, D>

pub fn sim(self) -> SimFlow<'a>

Available on crate feature sim only.

Creates a simulation for this builder, which can be used to run deterministic simulations of the Hydro program.

pub fn into_deploy<D: Deploy<'a>>(self) -> DeployFlow<'a, D>

pub fn with_process<P, D: Deploy<'a>>( self, process: &Process<'_, P>, spec: impl IntoProcessSpec<'a, D>, ) -> DeployFlow<'a, D>

pub fn with_remaining_processes<D: Deploy<'a>, S: IntoProcessSpec<'a, D> + 'a>( self, spec: impl Fn() -> S, ) -> DeployFlow<'a, D>

pub fn with_external<P, D: Deploy<'a>>( self, process: &External<'_, P>, spec: impl ExternalSpec<'a, D>, ) -> DeployFlow<'a, D>

pub fn with_remaining_externals<D: Deploy<'a>, S: ExternalSpec<'a, D> + 'a>( self, spec: impl Fn() -> S, ) -> DeployFlow<'a, D>

pub fn with_cluster<C, D: Deploy<'a>>( self, cluster: &Cluster<'_, C>, spec: impl ClusterSpec<'a, D>, ) -> DeployFlow<'a, D>

pub fn with_remaining_clusters<D: Deploy<'a>, S: ClusterSpec<'a, D> + 'a>( self, spec: impl Fn() -> S, ) -> DeployFlow<'a, D>

pub fn compile<D: Deploy<'a>>(self) -> CompiledFlow<'a>

pub fn deploy<D: Deploy<'a>>( self, env: &mut D::InstantiateEnv, ) -> DeployResult<'a, D>

pub fn generate_all_files( &self, prefix: &str, show_metadata: bool, show_location_groups: bool, use_short_labels: bool, ) -> Result<(), Box<dyn Error>>

Available on crate feature viz only.

pub fn generate_graph_with_config( &self, config: &GraphConfig, message_handler: Option<&dyn Fn(&str)>, ) -> Result<(), Box<dyn Error>>

Available on crate feature viz only.

pub fn generate_all_files_with_config( &self, config: &GraphConfig, prefix: &str, ) -> Result<(), Box<dyn Error>>

Available on crate feature viz only.