hydro_lang::live_collections::boundedness

Trait Boundedness

pub trait Boundedness: KeyedBoundFoldLike + Sealed {
    const BOUNDED: bool;
    const BOUND_KIND: BoundKind = _;
}

A marker trait indicating whether a stream’s length is bounded (finite) or unbounded (potentially infinite).

Implementors of this trait use it to signal the boundedness property of a stream.

Required Associated Constants

const BOUNDED: bool

true if the bound is Bounded, false if it is Unbounded.

Provided Associated Constants

const BOUND_KIND: BoundKind = _

The BoundKind corresponding to this type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.