hydro_lang::live_collections::keyed_singleton

Trait KeyedSingletonBound

pub trait KeyedSingletonBound {
    type UnderlyingBound: Boundedness;
    type ValueBound: Boundedness;
}

A marker trait indicating which components of a KeyedSingleton may change.

In addition to Bounded (all entries are fixed) and Unbounded (entries may be added / removed / changed), this also includes an additional variant BoundedValue, which indicates that entries may be added over time, but once an entry is added it will never be removed and its value will never change.

Required Associated Types

type UnderlyingBound: Boundedness

The Boundedness of the Stream underlying the keyed singleton.

type ValueBound: Boundedness

The Boundedness of each entry’s value; Bounded means it is immutable.