hydro_lang::live_collections::sliced

Trait Slicable

pub trait Slicable<'a, L: Location<'a>> {
    type Slice;
    type Backtrace;

    // Required methods
    fn preferred_tick(&self) -> Option<Tick<L>>;
    fn get_location(&self) -> &L;
    fn slice(self, tick: &Tick<L>, backtrace: Self::Backtrace) -> Self::Slice;

    // Provided method
    fn create_tick(&self) -> Tick<L>
       where L: NoTick { ... }
}

A trait for live collections which can be sliced into bounded versions at a tick.

Required Associated Types

type Slice

The sliced version of this live collection.

type Backtrace

The type of backtrace associated with this slice.

Required Methods

fn preferred_tick(&self) -> Option<Tick<L>>

Gets the preferred tick to slice at. Used for atomic slicing.

fn get_location(&self) -> &L

Gets the location associated with this live collection.

fn slice(self, tick: &Tick<L>, backtrace: Self::Backtrace) -> Self::Slice

Slices this live collection at the given tick.

Non-Determinism

Slicing a live collection may involve non-determinism, such as choosing which messages to include in a batch.

Provided Methods

fn create_tick(&self) -> Tick<L>
where L: NoTick,

Creates a tick that is appropriate for the collection’s location.