hydro_lang::location

Trait Location

pub trait Location<'a>: DynLocation {
    type Root: Location<'a>;

    // Required method
    fn root(&self) -> Self::Root;

    // Provided methods
    fn try_tick(&self) -> Option<Tick<Self>> { ... }
    fn id(&self) -> LocationId { ... }
    fn tick(&self) -> Tick<Self>
       where Self: NoTick { ... }
    fn spin(&self) -> Stream<(), Self, Unbounded, TotalOrder, ExactlyOnce>
       where Self: Sized + NoTick { ... }
    fn source_stream<T, E>(
        &self,
        e: impl QuotedWithContext<'a, E, Self>,
    ) -> Stream<T, Self, Unbounded, TotalOrder, ExactlyOnce>
       where E: FuturesStream<Item = T> + Unpin,
             Self: Sized + NoTick { ... }
    fn source_iter<T, E>(
        &self,
        e: impl QuotedWithContext<'a, E, Self>,
    ) -> Stream<T, Self, Unbounded, TotalOrder, ExactlyOnce>
       where E: IntoIterator<Item = T>,
             Self: Sized + NoTick { ... }
    fn source_cluster_members<C: 'a>(
        &self,
        cluster: &Cluster<'a, C>,
    ) -> KeyedStream<MemberId<C>, MembershipEvent, Self, Unbounded>
       where Self: Sized + NoTick { ... }
    fn source_external_bytes<L>(
        &self,
        from: &External<'_, L>,
    ) -> (ExternalBytesPort, Stream<BytesMut, Self, Unbounded, TotalOrder, ExactlyOnce>)
       where Self: Sized + NoTick { ... }
    fn source_external_bincode<L, T, O: Ordering, R: Retries>(
        &self,
        from: &External<'_, L>,
    ) -> (ExternalBincodeSink<T, NotMany, O, R>, Stream<T, Self, Unbounded, O, R>)
       where Self: Sized + NoTick,
             T: Serialize + DeserializeOwned { ... }
    fn sim_input<T, O: Ordering, R: Retries>(
        &self,
    ) -> (SimSender<T, O, R>, Stream<T, Self, Unbounded, O, R>)
       where Self: Sized + NoTick,
             T: Serialize + DeserializeOwned { ... }
    fn bind_single_client<L, T, Codec: Encoder<T> + Decoder>(
        &self,
        from: &External<'_, L>,
        port_hint: NetworkHint,
    ) -> (ExternalBytesPort<NotMany>, Stream<<Codec as Decoder>::Item, Self, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, Stream<T, Self, Unbounded, TotalOrder, ExactlyOnce>>)
       where Self: Sized + NoTick { ... }
    fn bind_single_client_bincode<L, InT: DeserializeOwned, OutT: Serialize>(
        &self,
        from: &External<'_, L>,
    ) -> (ExternalBincodeBidi<InT, OutT, NotMany>, Stream<InT, Self, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, Stream<OutT, Self, Unbounded, TotalOrder, ExactlyOnce>>)
       where Self: Sized + NoTick { ... }
    fn bidi_external_many_bytes<L, T, Codec: Encoder<T> + Decoder>(
        &self,
        from: &External<'_, L>,
        port_hint: NetworkHint,
    ) -> (ExternalBytesPort<Many>, KeyedStream<u64, <Codec as Decoder>::Item, Self, Unbounded, TotalOrder, ExactlyOnce>, KeyedStream<u64, MembershipEvent, Self, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, KeyedStream<u64, T, Self, Unbounded, NoOrder, ExactlyOnce>>)
       where Self: Sized + NoTick { ... }
    fn bidi_external_many_bincode<L, InT: DeserializeOwned, OutT: Serialize>(
        &self,
        from: &External<'_, L>,
    ) -> (ExternalBincodeBidi<InT, OutT, Many>, KeyedStream<u64, InT, Self, Unbounded, TotalOrder, ExactlyOnce>, KeyedStream<u64, MembershipEvent, Self, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, KeyedStream<u64, OutT, Self, Unbounded, NoOrder, ExactlyOnce>>)
       where Self: Sized + NoTick { ... }
    fn singleton<T>(
        &self,
        e: impl QuotedWithContext<'a, T, Self>,
    ) -> Singleton<T, Self, Unbounded>
       where T: Clone,
             Self: Sized { ... }
    fn source_interval(
        &self,
        interval: impl QuotedWithContext<'a, Duration, Self> + Copy + 'a,
        _nondet: NonDet,
    ) -> Stream<Instant, Self, Unbounded, TotalOrder, ExactlyOnce>
       where Self: Sized + NoTick { ... }
    fn source_interval_delayed(
        &self,
        delay: impl QuotedWithContext<'a, Duration, Self> + Copy + 'a,
        interval: impl QuotedWithContext<'a, Duration, Self> + Copy + 'a,
        _nondet: NonDet,
    ) -> Stream<Instant, Self, Unbounded, TotalOrder, ExactlyOnce>
       where Self: Sized + NoTick { ... }
    fn forward_ref<S>(&self) -> (ForwardHandle<'a, S>, S)
       where S: CycleCollection<'a, ForwardRef, Location = Self>,
             Self: NoTick { ... }

}

Required Associated Types

type Root: Location<'a>

Required Methods

fn root(&self) -> Self::Root

Provided Methods

fn try_tick(&self) -> Option<Tick<Self>>

fn id(&self) -> LocationId

fn tick(&self) -> Tick<Self>
where Self: NoTick,

fn spin(&self) -> Stream<(), Self, Unbounded, TotalOrder, ExactlyOnce>
where Self: Sized + NoTick,

fn source_stream<T, E>( &self, e: impl QuotedWithContext<'a, E, Self>, ) -> Stream<T, Self, Unbounded, TotalOrder, ExactlyOnce>
where E: FuturesStream<Item = T> + Unpin, Self: Sized + NoTick,

fn source_iter<T, E>( &self, e: impl QuotedWithContext<'a, E, Self>, ) -> Stream<T, Self, Unbounded, TotalOrder, ExactlyOnce>
where E: IntoIterator<Item = T>, Self: Sized + NoTick,

fn source_cluster_members<C: 'a>( &self, cluster: &Cluster<'a, C>, ) -> KeyedStream<MemberId<C>, MembershipEvent, Self, Unbounded>
where Self: Sized + NoTick,

fn source_external_bytes<L>( &self, from: &External<'_, L>, ) -> (ExternalBytesPort, Stream<BytesMut, Self, Unbounded, TotalOrder, ExactlyOnce>)
where Self: Sized + NoTick,

fn source_external_bincode<L, T, O: Ordering, R: Retries>( &self, from: &External<'_, L>, ) -> (ExternalBincodeSink<T, NotMany, O, R>, Stream<T, Self, Unbounded, O, R>)
where Self: Sized + NoTick, T: Serialize + DeserializeOwned,

fn sim_input<T, O: Ordering, R: Retries>( &self, ) -> (SimSender<T, O, R>, Stream<T, Self, Unbounded, O, R>)
where Self: Sized + NoTick, T: Serialize + DeserializeOwned,

Available on crate feature sim only.

Sets up a simulated input port on this location, returning a handle to send messages to the location as well as a stream of received messages.

fn bind_single_client<L, T, Codec: Encoder<T> + Decoder>( &self, from: &External<'_, L>, port_hint: NetworkHint, ) -> (ExternalBytesPort<NotMany>, Stream<<Codec as Decoder>::Item, Self, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, Stream<T, Self, Unbounded, TotalOrder, ExactlyOnce>>)
where Self: Sized + NoTick,

Establishes a server on this location to receive a bidirectional connection from a single client, identified by the given External handle. Returns a port handle for the external process to connect to, a stream of incoming messages, and a handle to send outgoing messages.

Example

let node = flow.process::<()>();
let external = flow.external::<()>();
let (port, incoming, outgoing) =
    node.bind_single_client::<_, Bytes, LengthDelimitedCodec>(&external, NetworkHint::Auto);
outgoing.complete(incoming.map(q!(|data /* : Bytes */| {
    let mut resp: Vec<u8> = data.into();
    resp.push(42);
    resp.into() // : Bytes
})));

let nodes = flow // ... with_process and with_external

deployment.deploy().await.unwrap();
deployment.start().await.unwrap();

let (mut external_out, mut external_in) = nodes.connect(port).await;
external_in.send(vec![1, 2, 3].into()).await.unwrap();
assert_eq!(
    external_out.next().await.unwrap().unwrap(),
    vec![1, 2, 3, 42]
);

fn bind_single_client_bincode<L, InT: DeserializeOwned, OutT: Serialize>( &self, from: &External<'_, L>, ) -> (ExternalBincodeBidi<InT, OutT, NotMany>, Stream<InT, Self, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, Stream<OutT, Self, Unbounded, TotalOrder, ExactlyOnce>>)
where Self: Sized + NoTick,

fn bidi_external_many_bytes<L, T, Codec: Encoder<T> + Decoder>( &self, from: &External<'_, L>, port_hint: NetworkHint, ) -> (ExternalBytesPort<Many>, KeyedStream<u64, <Codec as Decoder>::Item, Self, Unbounded, TotalOrder, ExactlyOnce>, KeyedStream<u64, MembershipEvent, Self, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, KeyedStream<u64, T, Self, Unbounded, NoOrder, ExactlyOnce>>)
where Self: Sized + NoTick,

fn bidi_external_many_bincode<L, InT: DeserializeOwned, OutT: Serialize>( &self, from: &External<'_, L>, ) -> (ExternalBincodeBidi<InT, OutT, Many>, KeyedStream<u64, InT, Self, Unbounded, TotalOrder, ExactlyOnce>, KeyedStream<u64, MembershipEvent, Self, Unbounded, TotalOrder, ExactlyOnce>, ForwardHandle<'a, KeyedStream<u64, OutT, Self, Unbounded, NoOrder, ExactlyOnce>>)
where Self: Sized + NoTick,

fn singleton<T>( &self, e: impl QuotedWithContext<'a, T, Self>, ) -> Singleton<T, Self, Unbounded>
where T: Clone, Self: Sized,

Constructs a Singleton materialized at this location with the given static value.

Example

let tick = process.tick();
let singleton = tick.singleton(q!(5));
// 5

fn source_interval( &self, interval: impl QuotedWithContext<'a, Duration, Self> + Copy + 'a, _nondet: NonDet, ) -> Stream<Instant, Self, Unbounded, TotalOrder, ExactlyOnce>
where Self: Sized + NoTick,

Generates a stream with values emitted at a fixed interval, with each value being the current time (as an [tokio::time::Instant]).

The clock source used is monotonic, so elements will be emitted in increasing order.

Non-Determinism

Because this stream is generated by an OS timer, it will be non-deterministic because each timestamp will be arbitrary.

fn source_interval_delayed( &self, delay: impl QuotedWithContext<'a, Duration, Self> + Copy + 'a, interval: impl QuotedWithContext<'a, Duration, Self> + Copy + 'a, _nondet: NonDet, ) -> Stream<Instant, Self, Unbounded, TotalOrder, ExactlyOnce>
where Self: Sized + NoTick,

Generates a stream with values emitted at a fixed interval (with an initial delay), with each value being the current time (as an [tokio::time::Instant]).

The clock source used is monotonic, so elements will be emitted in increasing order.

Non-Determinism

Because this stream is generated by an OS timer, it will be non-deterministic because each timestamp will be arbitrary.

fn forward_ref<S>(&self) -> (ForwardHandle<'a, S>, S)
where S: CycleCollection<'a, ForwardRef, Location = Self>, Self: NoTick,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.