hydro_lang::sim::compiled

Struct CompiledSimInstance

pub struct CompiledSimInstance<'a> { /* private fields */ }

Available on crate feature sim and stageleft_runtime only.

A single instance of a compiled Hydro simulation, which provides methods to interactively execute the simulation, feed inputs, and receive outputs.

Implementations

impl<'a> CompiledSimInstance<'a>

pub fn connect_sink_bincode<T: Serialize + 'static, M, O: Ordering, R: Retries>( &mut self, port: &ExternalBincodeSink<T, M, O, R>, ) -> SimSender<T, O, R>

๐Deprecated: Use connect instead

Like the corresponding method on crate::compile::deploy::DeployResult, connects to the given input port, and returns a closure that can be used to send messages to it.

pub fn connect_source_bincode<T: DeserializeOwned + 'static, O: Ordering, R: Retries>( &mut self, port: &ExternalBincodeStream<T, O, R>, ) -> SimReceiver<'a, T, O, R>

๐Deprecated: Use connect instead

Like the corresponding method on crate::compile::deploy::DeployResult, connects to the given output port, and returns a stream that can be used to receive messages from it.

pub fn connect<'b, P: ConnectableAsync<&'b mut Self>>( &'b mut self, port: P, ) -> <P as ConnectableAsync<&'b mut Self>>::Output

Establishes a connection to the given input or output port, returning either a SimSender (for input ports) or a stream (for output ports). This should be invoked before calling Self::launch, and should only be invoked once per port.

pub fn launch(self)

Launches the simulation, which will asynchronously simulate the Hydro program. This should be invoked after connecting all inputs and outputs, but before receiving any messages.

pub fn schedule_with_logger<W: Write>( self, log_writer: W, ) -> impl use<W> + Future<Output = ()>

Returns a future that schedules simulation with the given logger for reporting the simulation trace.

See Self::launch for more details.